added the components to the page
